sql: fix overflow for soft limits
We're tracking `LimitHint` property as a float and later convert it to
an int64. Previously, this could lead to overflows when the limit hint
value exceeds the range of int64, and we now instead will use a value of
0 (i.e. disable the soft limit).

Release note (bug fix): Previously, CockroachDB could encounter an
internal error when evaluating queries with OFFSET and LIMIT clauses
when the addition of the `offset` and the `limit` value would be larger
than `int64` range.

+++ pkg/sql/index_join.go
@@ -37,7 +37,7 @@ type indexJoinNode struct {
 
 	reqOrdering ReqOrdering
 
-	limitHint int
+	limitHint int64
 }
 
 func (n *indexJoinNode) startExec(params runParams) error {
